fix: build warnings, test grep bug, mandate keywords in challenge summary
- Add mandate_keywords to buildChallengeSummary() so the challenge
  preamble reinforces the agent's stated objective (not just task context)
- Fix 5 build warnings: float-to-int cast, sign-conversion casts,
  unused variable moved inside #ifdef
- Fix grep -c || echo "0" pattern in test_d1_reconciliation.sh that
  produced "0\n0" on zero matches, breaking integer comparison
- Add demo and test files: fan-out demo, sandbox stress test,
  naab-44 quick-build config and script

+++ src/interpreter/debugger.cpp
@@ -107,7 +107,7 @@ bool InterpreterDebugger::shouldBreak(const ast::ASTNode& node) const {
 
     auto it = breakpoints_.find(loc.filename);
     if (it != breakpoints_.end()) {
-        return it->second.count(loc.line) > 0;
+        return it->second.count(static_cast<size_t>(loc.line)) > 0;
     }
 
     return false;
